backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.1
pydantic-settings==2.4.0
python-jose==3.3.0
bcrypt==3.2.2
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
python-multipart==0.0.9
flake8
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-multipart
python-dotenv
asyncpg
httpx
pytest
pytest-asyncio
flake8
